+++ R/seq.R
@@ -1,18 +1,18 @@
-#' Create an evenly spaced sequence of values from lowest to highest.
+#' Create the full sequence of values in a vector.
 #'
-#' @param x A numeric vector
-#' @param n Number of values
+#' This is useful if you want to fill in missing values that should have
+#' been observed but weren't. For example, \code{full_seq(c(1, 2, 4, 6), 1)}
+#' will return \code{1:6}.
+#'
+#' @param x A numeric vector.
+#' @param period Gap between each observation. The existing data will be
+#'   checked to ensure that it is actually of this periodicity.
+#' @param tol Numerical tolerance for checking periodicity.
 #' @export
 #' @examples
-#' seq_range(1:100, 5)
-seq_range <- function(x, n) {
-  rng <- range(x, na.rm = TRUE)
-  seq(rng[1], rng[2], length = n)
-}
-
-
-seq_time <- function(x, period) {
-  if (any(x %% period > 1e-6)) {
+#' full_seq(c(1, 2, 4, 5, 10), 1)
+full_seq <- function(x, period, tol = 1e-6) {
+  if (any(x %% period > tol)) {
     stop("`x` is not a regular sequence.", call. = FALSE)
   }